Move benchmark test to right folder
Make get_warnings_list a fixture.
