+++ tests/benchmarks/__snapshots__sphinx_lt_8/test_schema_benchmark.ambr
